Refactored tasks for old proving pipeline
