+++ packages/sequencer/src/index.ts
@@ -24,12 +24,20 @@ export * from "./protocol/baselayer/MinaBaseLayer";
 export * from "./protocol/baselayer/NoopBaseLayer";
 export * from "./protocol/production/helpers/UntypedOption";
 export * from "./protocol/production/helpers/UntypedStateTransition";
-export * from "./protocol/production/tasks/BlockProvingTask";
+export * from "./protocol/production/tasks/TransactionProvingTask";
 export * from "./protocol/production/tasks/RuntimeProvingTask";
-export * from "./protocol/production/tasks/RuntimeTaskParameters";
 export * from "./protocol/production/tasks/StateTransitionTask";
-export * from "./protocol/production/tasks/StateTransitionTaskParameters";
+export * from "./protocol/production/tasks/StateTransitionReductionTask";
 export * from "./protocol/production/tasks/NewBlockTask";
+export * from "./protocol/production/tasks/serializers/ArtifactionRecordSerializer";
+export * from "./protocol/production/tasks/serializers/BlockProofSerializer";
+export * from "./protocol/production/tasks/serializers/DecodedStateSerializer";
+export * from "./protocol/production/tasks/serializers/NewBlockProvingParametersSerializer";
+export * from "./protocol/production/tasks/serializers/RuntimeProofParametersSerializer";
+export * from "./protocol/production/tasks/serializers/RuntimeVerificationKeyAttestationSerializer";
+export * from "./protocol/production/tasks/serializers/StateTransitionParametersSerializer";
+export * from "./protocol/production/tasks/serializers/TransactionProvingTaskParameterSerializer";
+export * from "./protocol/production/tasks/serializers/VerificationKeySerializer";
 export * from "./protocol/production/trigger/BlockTrigger";
 export * from "./protocol/production/trigger/ManualBlockTrigger";
 export * from "./protocol/production/trigger/TimedBlockTrigger";
