+++ packages/sequencer/src/protocol/production/BatchProducerModule.ts
@@ -23,13 +23,13 @@ import { AsyncMerkleTreeStore } from "../../state/async/AsyncMerkleTreeStore";
 import { BlockResult, BlockWithResult } from "../../storage/model/Block";
 import { VerificationKeyService } from "../runtime/RuntimeVerificationKeyService";
 
-import { BlockProverParameters } from "./tasks/BlockProvingTask";
-import { StateTransitionProofParameters } from "./tasks/StateTransitionTaskParameters";
-import { RuntimeProofParameters } from "./tasks/RuntimeTaskParameters";
 import { TransactionTraceService } from "./TransactionTraceService";
 import { BlockTaskFlowService } from "./BlockTaskFlowService";
 import { NewBlockProverParameters } from "./tasks/NewBlockTask";
-import { BlockProofSerializer } from "./helpers/BlockProofSerializer";
+import { BlockProofSerializer } from "./tasks/serializers/BlockProofSerializer";
+import { RuntimeProofParameters } from "./tasks/RuntimeProvingTask";
+import { StateTransitionProofParameters } from "./tasks/StateTransitionTask";
+import { BlockProverParameters } from "./tasks/TransactionProvingTask";
 
 export type StateRecord = Record<string, Field[] | undefined>;
