+++ packages/sequencer/src/protocol/production/tasks/BlockReductionTask.ts
@@ -1,0 +1,67 @@
+import { inject, injectable, Lifecycle, scoped } from "tsyringe";
+import {
+  BlockProof,
+  BlockProvable,
+  MandatoryProtocolModulesRecord,
+  Protocol,
+  ProtocolModulesRecord,
+} from "@proto-kit/protocol";
+import {
+  CompileRegistry,
+  ProvableMethodExecutionContext,
+} from "@proto-kit/common";
+
+import { TaskWorkerModule } from "../../../worker/worker/TaskWorkerModule";
+import { Task, TaskSerializer } from "../../../worker/flow/Task";
+import {
+  PairProofTaskSerializer,
+  PairTuple,
+  ProofTaskSerializer,
+} from "../../../helpers/utils";
+import { VerificationKeyService } from "../../runtime/RuntimeVerificationKeyService";
+
+@injectable()
+@scoped(Lifecycle.ContainerScoped)
+export class BlockReductionTask
+  extends TaskWorkerModule
+  implements Task<PairTuple<BlockProof>, BlockProof>
+{
+  private readonly blockProver: BlockProvable;
+
+  public name = "blockReduction";
+
+  public constructor(
+    @inject("Protocol")
+    private readonly protocol: Protocol<
+      MandatoryProtocolModulesRecord & ProtocolModulesRecord
+    >,
+    private readonly executionContext: ProvableMethodExecutionContext,
+    private readonly compileRegistry: CompileRegistry,
+    private readonly verificationKeyService: VerificationKeyService
+  ) {
+    super();
+    this.blockProver = this.protocol.blockProver;
+  }
+
+  public inputSerializer(): TaskSerializer<PairTuple<BlockProof>> {
+    return new PairProofTaskSerializer(
+      this.blockProver.zkProgrammable.zkProgram[0].Proof
+    );
+  }
+
+  public resultSerializer(): TaskSerializer<BlockProof> {
+    return new ProofTaskSerializer(
+      this.blockProver.zkProgrammable.zkProgram[0].Proof
+    );
+  }
+
+  public async compute(input: PairTuple<BlockProof>): Promise<BlockProof> {
+    const [r1, r2] = input;
+    await this.blockProver.merge(r1.publicInput, r1, r2);
+    return await this.executionContext.current().result.prove<BlockProof>();
+  }
+
+  public async prepare(): Promise<void> {
+    await this.blockProver.compile(this.compileRegistry);
+  }
+}
