+++ packages/sequencer/src/protocol/production/tasks/CircuitCompilerTask.ts
@@ -19,51 +19,17 @@ import {
 
 import { TaskSerializer } from "../../../worker/flow/Task";
 import { UnpreparingTask } from "../../../worker/flow/UnpreparingTask";
-import { VerificationKeySerializer } from "../helpers/VerificationKeySerializer";
+
+import {
+  ArtifactRecordSerializer,
+  SerializedArtifactRecord,
+} from "./serializers/ArtifactionRecordSerializer";
 
 export type CompilerTaskParams = {
   existingArtifacts: ArtifactRecord;
   targets: string[];
   runtimeVKRoot?: string;
 };
-
-export type SerializedArtifactRecord = Record<
-  string,
-  { verificationKey: { hash: string; data: string } }
->;
-
-export class ArtifactRecordSerializer {
-  public toJSON(input: ArtifactRecord): SerializedArtifactRecord {
-    const temp: SerializedArtifactRecord = Object.keys(
-      input
-    ).reduce<SerializedArtifactRecord>((accum, key) => {
-      return {
-        ...accum,
-        [key]: {
-          verificationKey: VerificationKeySerializer.toJSON(
-            input[key].verificationKey
-          ),
-        },
-      };
-    }, {});
-    return temp;
-  }
-
-  public fromJSON(json: SerializedArtifactRecord): ArtifactRecord {
-    if (json === undefined || json === null) return {};
-
-    return Object.keys(json).reduce<ArtifactRecord>((accum, key) => {
-      return {
-        ...accum,
-        [key]: {
-          verificationKey: VerificationKeySerializer.fromJSON(
-            json[key].verificationKey
-          ),
-        },
-      };
-    }, {});
-  }
-}
 
 @injectable()
 @scoped(Lifecycle.ContainerScoped)
