+++ packages/sequencer/src/protocol/production/tasks/serializers/VerificationKeySerializer.ts
